test_2

+++ app/Services/RouteService.php
@@ -4,6 +4,7 @@ namespace App\Services;
 
 use App\Interfaces\RouterInterface;
 use App\Helpers\Waypoints;
+use Illuminate\Support\Facades\Gate;
 
 class RouteService
 {
@@ -32,6 +33,7 @@ class RouteService
 
     public function create(array $data)
     {
+        Gate::authorize('manage-routes');
         if (isset($data['waypoints'])) {
             $data['waypoints'] = Waypoints::normalize($data['waypoints']);
         }
@@ -41,6 +43,7 @@ class RouteService
 
     public function delete($id): bool
     {
+        Gate::authorize('manage-routes');
         return (bool) $this->repo->delete($id);
     }
 }
