+++ app/Services/ServiceTypeService.php
@@ -3,6 +3,7 @@
 namespace App\Services;
 
 use App\Interfaces\ServiceTypeInterface;
+use Illuminate\Support\Facades\Gate;
 
 class ServiceTypeService
 {
@@ -25,16 +26,19 @@ class ServiceTypeService
 
     public function create(array $data)
     {
+        Gate::authorize('manage-services');
         return $this->repo->createServiceType($data);
     }
 
     public function update($id, array $data)
     {
+        Gate::authorize('manage-services');
         return $this->repo->updateServiceType($id, $data);
     }
 
     public function delete($id)
     {
+        Gate::authorize('manage-services');
         return $this->repo->deleteServiceType($id);
     }
 }
